requirements.txt
Django==1.11.4
pyaml==17.8.0